pyproject.toml
[tool.poetry]
name = "goon"
version = "0.1.0"
description = ""
authors = ["DaniFdz <danifernandezzzzzz@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
pillow = "^10.1.0"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"
daphne = "^4.0.0"
python-dotenv = "^1.0.0"
mysql-connector-python = "^8.2.0"
pymysql = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-xdist = "^3.5.0"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
model-bakery = "^1.17.0"
coverage = "^7.3.4"
pre-commit = "^3.6.0"
colorlog = "^6.8.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["core"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"
pythonpath = "."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
